feat: 라운지 메인, 라운지 상세 페이지 ssr 적용
